Migrated PdfTemplating.XslFO.Razor.AspNetMvc.csproj to use 'PackageReferences' and new project package properties for Nuget package management.  This should now enable successful publish to Nuget by GitHub actions using the latest project based capabilities for Nuget package properties and references.
This was done by enabling the latest package management configuration within .csproj instead of packages.config file, and updated project file to also contain Nuget properties within project file for use by NuGet.Build.Tasks.Pack package -- which provides a Pack Target for the command line 'donet pack ...'. We now eliminated the need for NuSpec file but added the complexity of having to directly edit the .csproj file to update (which for .Net v4.8 is not friendly in the Visual Studio).